+++ Rebus/Config/RebusConfigurer.cs
@@ -161,15 +161,15 @@ namespace Rebus.Config
 
             PossiblyRegisterDefault<IBackoffStrategy>(c => new SimpleConstantPollingBackoffStrategy());
 
-            PossiblyRegisterDefault<IWorkerFactory>(c =>
-            {
-                var transport = c.Get<ITransport>();
-                var pipeline = c.Get<IPipeline>();
-                var pipelineInvoker = c.Get<IPipelineInvoker>();
-                var backoffStrategy = c.Get<IBackoffStrategy>();
-                var rebusLoggerFactory = c.Get<IRebusLoggerFactory>();
-                return new ThreadWorkerFactory(transport, pipeline, pipelineInvoker, backoffStrategy, rebusLoggerFactory, _options, c.Get<RebusBus>);
-            });
+            //PossiblyRegisterDefault<IWorkerFactory>(c =>
+            //{
+            //    var transport = c.Get<ITransport>();
+            //    var pipeline = c.Get<IPipeline>();
+            //    var pipelineInvoker = c.Get<IPipelineInvoker>();
+            //    var backoffStrategy = c.Get<IBackoffStrategy>();
+            //    var rebusLoggerFactory = c.Get<IRebusLoggerFactory>();
+            //    return new ThreadWorkerFactory(transport, pipeline, pipelineInvoker, backoffStrategy, rebusLoggerFactory, _options, c.Get<RebusBus>);
+            //});
 
             PossiblyRegisterDefault<ISyncBackoffStrategy>(c =>
             {
@@ -185,17 +185,17 @@ namespace Rebus.Config
                 return new DefaultSyncBackoffStrategy(backoffTimes.SelectMany(e => e));
             });
 
-            //PossiblyRegisterDefault<IWorkerFactory>(c =>
-            //{
-            //    var transport = c.Get<ITransport>();
-            //    var rebusLoggerFactory = c.Get<IRebusLoggerFactory>();
-            //    var pipeline = c.Get<IPipeline>();
-            //    var pipelineInvoker = c.Get<IPipelineInvoker>();
-            //    var options = c.Get<Options>();
-            //    var busLifetimeEvents = c.Get<BusLifetimeEvents>();
-            //    var backoffStrategy = c.Get<ISyncBackoffStrategy>();
-            //    return new ThreadPoolWorkerFactory(transport, rebusLoggerFactory, pipeline, pipelineInvoker, options, c.Get<RebusBus>, busLifetimeEvents, backoffStrategy);
-            //});
+            PossiblyRegisterDefault<IWorkerFactory>(c =>
+            {
+                var transport = c.Get<ITransport>();
+                var rebusLoggerFactory = c.Get<IRebusLoggerFactory>();
+                var pipeline = c.Get<IPipeline>();
+                var pipelineInvoker = c.Get<IPipelineInvoker>();
+                var options = c.Get<Options>();
+                var busLifetimeEvents = c.Get<BusLifetimeEvents>();
+                var backoffStrategy = c.Get<ISyncBackoffStrategy>();
+                return new ThreadPoolWorkerFactory(transport, rebusLoggerFactory, pipeline, pipelineInvoker, options, c.Get<RebusBus>, busLifetimeEvents, backoffStrategy);
+            });
 
             PossiblyRegisterDefault<IErrorTracker>(c =>
             {