+++ CHANGELOG.md
@@ -1035,6 +1035,7 @@
 * Mark assemblies as CLS compliant becase VB.NET and F# programmers are most welcome too - thanks [NKnusperer]
 * Update Serilog dependency to 2.1.0
 * Limit number of workers to match max parallelism
+* Make thread pool-based workers default (old strategy can still be had by calling `o.UseClassicRebusWorkersMessageDispatch()`)
 
 ---